refactor: Rename to new_prompt() and :ZkNewPrompt / Move new_prompt() to zk.lua
