+++ lua/zk/util.lua
@@ -234,132 +234,4 @@ function M.get_dirs(cwd, depth, ignores)
   return dirs
 end
 
----Prompt for group, paths, templates, and directories for new note.
----@param cwd string?
----@param cb fun(ret: table?, config: table?)
-function M.prompt_new(cwd, cb)
-  local ret = {}
-  cwd = cwd or require("zk.util").notebook_root(vim.fn.getcwd())
-  if not cwd then
-    return
-  end
-  local config = M.get_config_toml(cwd) or {}
-  local groups = config.group -- Note that it is named `group` instead of `groups` in config.toml
-
-  ---@param a string
-  ---@param b string
-  ---@return boolean
-  local function sorter(a, b)
-    return a < b
-  end
-
-  ---@param group_names string[]
-  ---@param on_select function
-  function select_group(group_names, on_select)
-    table.sort(group_names, sorter)
-    vim.ui.select(group_names, { prompt = "Select a group" }, function(group_name)
-      if not group_name then
-        return
-      end
-      ret.group = group_name
-      on_select()
-    end)
-  end
-
-  ---@param on_select function
-  function select_directory_fs(on_select)
-    local dirs = M.get_dirs(cwd) or {}
-    table.insert(dirs, "")
-    table.sort(dirs, sorter)
-    vim.ui.select(dirs, { prompt = "Select a directory" }, function(dir)
-      if not dir then
-        return
-      end
-      ret.dir = dir
-      on_select()
-    end)
-  end
-
-  ---@param paths string[]
-  ---@param on_select function
-  function select_paths(paths, on_select)
-    table.sort(paths, sorter)
-    if #paths > 1 then
-      vim.ui.select(paths, { prompt = "Select a path" }, function(path)
-        if not path then
-          return
-        end
-        ret.dir = path
-        ret.template = groups[ret.group].note and groups[ret.group].note.template
-        on_select()
-      end)
-    elseif #paths == 1 then -- Apply automatically if only one path
-      ret.dir = paths[1]
-      ret.template = groups[ret.group].note and groups[ret.group].note.template
-      on_select()
-    end
-  end
-
-  ---@param on_select function
-  function select_template_fs(on_select)
-    local templates = M.get_templates(cwd)
-    if not templates or vim.tbl_count(templates) == 0 then
-      local msg = "Cannot find any templates in `.zk/templates`"
-      vim.notify(msg, vim.log.levels.ERROR, { title = "zk-nvim" })
-      return
-    end
-
-    local template_names = {}
-    for _, template in pairs(templates) do
-      table.insert(template_names, template.name)
-    end
-    table.sort(template_names, sorter)
-    vim.ui.select(template_names, { prompt = "Select a template" }, function(template_name)
-      if not template_name then
-        return
-      end
-      ret.template = template_name
-      on_select()
-    end)
-  end
-
-  if groups then -- groups exists
-    local group_names = vim.tbl_keys(groups)
-    select_group(group_names, function()
-      local paths = groups[ret.group] and groups[ret.group].paths
-      if paths then -- paths exists
-        select_paths(paths, function()
-          local template_name = groups[ret.group].note and groups[ret.group].note.template
-          if template_name then
-            ret.template = template_name
-            cb(ret, config)
-          else
-            select_template_fs(function()
-              cb(ret, config)
-            end)
-          end
-        end)
-      else -- paths does not exist
-        select_directory_fs(function()
-          local template_name = groups[ret.group].note and groups[ret.group].note.template
-          if template_name then
-            ret.template = template_name
-            cb(ret, config)
-          else
-            select_template_fs(function()
-              cb(ret, config)
-            end)
-          end
-        end)
-      end
-    end)
-  else -- groups does not exist
-    select_directory_fs(function()
-      select_template_fs(function()
-        cb(ret, config)
-      end)
-    end)
-  end
-end
-
 return M
